consolidation of dupes from StripesHub into loginServices
